requirements.txt
numpy==2.4.3
pandas==3.0.1
scipy==1.17.1
scikit-learn==1.8.0
matplotlib==3.10.8
seaborn==0.13.2
jupyterlab==4.5.6
python-dotenv==1.2.2
prettytable==3.10.0
